refactor: remove old packages
