+++ iroh-api/Cargo.toml
@@ -20,10 +20,8 @@ config = "0.13.1"
 futures = "0.3.21"
 iroh-metrics = { version = "0.1.3", path = "../iroh-metrics", default-features = false, features = ["rpc-grpc"] }
 iroh-resolver = { version = "0.1.3", path = "../iroh-resolver" }
-# iroh-rpc-client = { version = "0.1.3", path = "../iroh-rpc-client", default-features = false }
-# iroh-rpc-types = { version = "0.1.3", path = "../iroh-rpc-types", default-features = false }
-iroh-rpc-client = { package = "iroh-qrpc-client", version = "0.2.0", path = "../iroh-qrpc-client", default-features = false }
-iroh-rpc-types = { package = "iroh-qrpc-types", version = "0.2.0", path = "../iroh-qrpc-types", default-features = false }
+iroh-rpc-client = { version = "0.1.3", path = "../iroh-rpc-client", default-features = false }
+iroh-rpc-types = { version = "0.1.3", path = "../iroh-rpc-types", default-features = false }
 iroh-util = { version = "0.1.3", path = "../iroh-util" }
 libp2p = "0.50"
 mockall = { version = "0.11.2", optional = true }
