+++ iroh-gateway/Cargo.toml
@@ -30,10 +30,8 @@ hyper = "0.14.19"
 iroh-car = { version = "0.1.3", path = "../iroh-car" }
 iroh-metrics = { version = "0.1.3", path = "../iroh-metrics", default-features = false, features = ["gateway"] }
 iroh-resolver = { version = "0.1.3", path = "../iroh-resolver" }
-# iroh-rpc-client = { version = "0.1.3", path = "../iroh-rpc-client", default-features = false }
-# iroh-rpc-types = { version = "0.1.3", path = "../iroh-rpc-types", default-features = false }
-iroh-rpc-client = { package = "iroh-qrpc-client", version = "0.2.0", path = "../iroh-qrpc-client", default-features = false }
-iroh-rpc-types = { package = "iroh-qrpc-types", version = "0.2.0", path = "../iroh-qrpc-types", default-features = false }
+iroh-rpc-client = { version = "0.1.3", path = "../iroh-rpc-client", default-features = false }
+iroh-rpc-types = { version = "0.1.3", path = "../iroh-rpc-types", default-features = false }
 iroh-util = { version = "0.1.3", path = "../iroh-util" }
 libp2p = { version = "0.50", default-features = false }
 mime = "0.3"
