+++ iroh-one/Cargo.toml
@@ -24,10 +24,8 @@ iroh-gateway = { version = "0.1.3", path = "../iroh-gateway" }
 iroh-metrics = { version = "0.1.3", path = "../iroh-metrics", default-features = false }
 iroh-p2p = { version = "0.1.3", path = "../iroh-p2p", default-features = false, features = ["rpc-mem"] }
 iroh-resolver = { version = "0.1.3", path = "../iroh-resolver" }
-# iroh-rpc-client = { version = "0.1.3", path = "../iroh-rpc-client", default-features = false }
-# iroh-rpc-types = { version = "0.1.3", path = "../iroh-rpc-types", default-features = false }
-iroh-rpc-client = { package = "iroh-qrpc-client", version = "0.2.0", path = "../iroh-qrpc-client", default-features = false }
-iroh-rpc-types = { package = "iroh-qrpc-types", version = "0.2.0", path = "../iroh-qrpc-types", default-features = false }
+iroh-rpc-client = { version = "0.1.3", path = "../iroh-rpc-client", default-features = false }
+iroh-rpc-types = { version = "0.1.3", path = "../iroh-rpc-types", default-features = false }
 iroh-store = { version = "0.1.3", path = "../iroh-store", default-features = false, features = ["rpc-mem"] }
 iroh-util = { version = "0.1.3", path = "../iroh-util" }
 reqwest = { version = "0.11", features = ["rustls-tls"], default-features = false }
