+++ iroh-p2p/Cargo.toml
@@ -23,10 +23,8 @@ futures-util = "0.3.21"
 git-version = "0.3.5"
 iroh-bitswap = { version = "0.1.3", path = "../iroh-bitswap" }
 iroh-metrics = { version = "0.1.3", path = "../iroh-metrics", default-features = false, features = ["bitswap", "p2p"] }
-# iroh-rpc-client = { version = "0.1.3", path = "../iroh-rpc-client", default-features = false }
-# iroh-rpc-types = { version = "0.1.3", path = "../iroh-rpc-types", default-features = false }
-iroh-rpc-client = { package = "iroh-qrpc-client", version = "0.2.0", path = "../iroh-qrpc-client", default-features = false }
-iroh-rpc-types = { package = "iroh-qrpc-types", version = "0.2.0", path = "../iroh-qrpc-types", default-features = false }
+iroh-rpc-client = { version = "0.1.3", path = "../iroh-rpc-client", default-features = false }
+iroh-rpc-types = { version = "0.1.3", path = "../iroh-rpc-types", default-features = false }
 iroh-util = { version = "0.1.3", path = "../iroh-util" }
 lazy_static = "1.4"
 lru = "0.8"
