--- iroh-qrpc-types/README.md
@@ -1,27 +1,0 @@
-# iroh-rpc-types
-
-This crate defines the protocol for use by the
-[iroh-rpc-client](https://github.com/n0-computer/iroh/tree/main/iroh-rpc-client),
-which is used for [iroh](https://github.com/n0-computer/iroh) services to
-communicate internally via RPC.
-
-It defines a set of messages and traits that specify how each message is to be
-processed.
-
-These types can be used with any transport protocol in
-[quic-rpc](https://github.com/n0-computer/quic-rpc).
-
-## License
-
-<sup>
-Licensed under either of <a href="LICENSE-APACHE">Apache License, Version
-2.0</a> or <a href="LICENSE-MIT">MIT license</a> at your option.
-</sup>
-
-<br/>
-
-<sub>
-Unless you explicitly state otherwise, any contribution intentionally submitted
-for inclusion in this crate by you, as defined in the Apache-2.0 license, shall
-be dual licensed as above, without any additional terms or conditions.
-</sub>
