+++ iroh-resolver/Cargo.toml
@@ -25,8 +25,7 @@ cid = "0.9"
 fastmurmur3 = "0.1.2"
 futures = "0.3.21"
 iroh-metrics = { version = "0.1.3", path = "../iroh-metrics", default-features = false, features = ["resolver", "gateway"] }
-# iroh-rpc-client = { version = "0.1.3", path = "../iroh-rpc-client", default-features = false }
-iroh-rpc-client = { package = "iroh-qrpc-client", version = "0.2.0", path = "../iroh-qrpc-client", default-features = false }
+iroh-rpc-client = { version = "0.1.3", path = "../iroh-rpc-client", default-features = false }
 iroh-util = { version = "0.1.3", path = "../iroh-util", default-features = false }
 libipld = "0.15.0"
 libp2p = { version = "0.50", default-features = false, features = ["serde"] }
@@ -48,8 +47,7 @@ fnv = "1.0.7"
 [dev-dependencies]
 criterion = { version = "0.4.0", features = ["async_tokio"] }
 iroh-car = { version = "0.1.3", path = "../iroh-car" }
-# iroh-rpc-types = { version = "0.1.3", path = "../iroh-rpc-types", default-features = false }
-iroh-rpc-types = { package = "iroh-qrpc-types", version = "0.2.0", path = "../iroh-qrpc-types", default-features = false }
+iroh-rpc-types = { version = "0.1.3", path = "../iroh-rpc-types", default-features = false }
 iroh-store = { version = "0.1.3", path = "../iroh-store", default-features = false }
 proptest = "1.0.0"
 rand = "0.8.5"
