+++ iroh-rpc-client/src/client.rs
@@ -1,14 +1,12 @@
 use std::sync::atomic::AtomicUsize;
 use std::sync::Arc;
 
+use super::config::Config;
+use super::gateway::GatewayClient;
+use super::network::P2pClient;
+use super::store::StoreClient;
 use anyhow::{Context, Result};
-#[cfg(feature = "grpc")]
 use futures::{Stream, StreamExt};
-
-use crate::config::Config;
-use crate::gateway::GatewayClient;
-use crate::network::P2pClient;
-use crate::store::StoreClient;
 
 #[derive(Debug, Clone)]
 pub struct Client {
@@ -23,6 +21,12 @@ pub struct Client {
 pub(crate) struct StoreLBClient {
     clients: Vec<StoreClient>,
     pos: Arc<AtomicUsize>,
+}
+
+impl Default for StoreLBClient {
+    fn default() -> Self {
+        Self::new()
+    }
 }
 
 impl StoreLBClient {
@@ -50,6 +54,12 @@ impl StoreLBClient {
 pub(crate) struct P2pLBClient {
     clients: Vec<P2pClient>,
     pos: Arc<AtomicUsize>,
+}
+
+impl Default for P2pLBClient {
+    fn default() -> Self {
+        Self::new()
+    }
 }
 
 impl P2pLBClient {
@@ -133,8 +143,7 @@ impl Client {
         self.store.get().context("missing rpc store connection")
     }
 
-    #[cfg(feature = "grpc")]
-    pub async fn check(&self) -> crate::status::StatusTable {
+    pub async fn check(&self) -> super::status::StatusTable {
         let g = if let Some(ref g) = self.gateway {
             Some(g.check().await)
         } else {
@@ -150,13 +159,12 @@ impl Client {
         } else {
             None
         };
-        crate::status::StatusTable::new(g, p, s)
+        super::status::StatusTable::new(g, p, s)
     }
 
-    #[cfg(feature = "grpc")]
-    pub async fn watch(self) -> impl Stream<Item = crate::status::StatusTable> {
+    pub async fn watch(self) -> impl Stream<Item = super::status::StatusTable> {
         async_stream::stream! {
-            let mut status_table: crate::status::StatusTable = Default::default();
+            let mut status_table: super::status::StatusTable = Default::default();
             let mut streams = Vec::new();
 
             if let Some(ref g) = self.gateway {
@@ -178,142 +186,5 @@ impl Client {
                 yield status_table.clone();
             }
         }
-    }
-}
-
-// TODO: write tests for mem transport
-#[cfg(all(test, feature = "grpc"))]
-mod tests {
-    use super::*;
-    use std::net::SocketAddr;
-
-    use iroh_rpc_types::test::test_server;
-    use tonic::transport::Server as TonicServer;
-    use tonic_health::{
-        server::{health_reporter, HealthReporter},
-        ServingStatus,
-    };
-
-    use crate::status::{ServiceStatus, StatusRow, StatusTable};
-    use crate::{gateway, network, store};
-
-    struct TestService {}
-    #[tonic::async_trait]
-    impl test_server::Test for TestService {}
-
-    async fn make_service(
-        name: &'static str,
-        addr: SocketAddr,
-    ) -> Result<(HealthReporter, tokio::task::JoinHandle<()>)> {
-        let (mut reporter, service) = health_reporter();
-        reporter
-            .set_serving::<test_server::TestServer<TestService>>()
-            .await;
-
-        let task = tokio::spawn(async move {
-            TonicServer::builder()
-                .add_service(service)
-                .serve(addr)
-                .await
-                .unwrap();
-        });
-        reporter
-            .set_service_status(name, ServingStatus::Serving)
-            .await;
-
-        Ok((reporter, task))
-    }
-
-    #[tokio::test]
-    async fn client_status() {
-        let cfg = Config::default_network();
-
-        let gateway_name = gateway::NAME;
-        let p2p_name = network::NAME;
-        let store_name = store::NAME;
-
-        // make the services with the expected service names
-        let (mut gateway_reporter, gateway_task) = make_service(
-            gateway::SERVICE_NAME,
-            cfg.gateway_addr
-                .as_ref()
-                .unwrap()
-                .try_as_socket_addr()
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-        let (mut p2p_reporter, p2p_task) = make_service(
-            network::SERVICE_NAME,
-            cfg.p2p_addr.as_ref().unwrap().try_as_socket_addr().unwrap(),
-        )
-        .await
-        .unwrap();
-        let (mut store_reporter, store_task) = make_service(
-            store::SERVICE_NAME,
-            cfg.store_addr
-                .as_ref()
-                .unwrap()
-                .try_as_socket_addr()
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-        let client = Client::new(cfg).await.unwrap();
-
-        // test `check`
-        // expect the names to be the hard-coded display names
-        let mut expect = StatusTable::new(
-            Some(StatusRow::new(gateway_name, 1, ServiceStatus::Serving)),
-            Some(StatusRow::new(p2p_name, 1, ServiceStatus::Serving)),
-            Some(StatusRow::new(store_name, 1, ServiceStatus::Serving)),
-        );
-        let mut got = client.check().await;
-        assert_eq!(expect, got);
-
-        // test `watch`
-        let stream = client.watch().await;
-        tokio::pin!(stream);
-
-        // each status gets reported for the first time in a non-deterministic order
-        stream.next().await.unwrap();
-        stream.next().await.unwrap();
-        got = stream.next().await.unwrap();
-
-        assert_eq!(expect, got);
-
-        // update gateway
-        expect
-            .update(StatusRow::new(gateway_name, 1, ServiceStatus::Unknown))
-            .unwrap();
-        gateway_reporter
-            .set_service_status(gateway::SERVICE_NAME, ServingStatus::Unknown)
-            .await;
-        let got = stream.next().await.unwrap();
-        assert_eq!(expect, got);
-
-        // update p2p
-        expect
-            .update(StatusRow::new(p2p_name, 1, ServiceStatus::NotServing))
-            .unwrap();
-        p2p_reporter
-            .set_service_status(network::SERVICE_NAME, ServingStatus::NotServing)
-            .await;
-        let got = stream.next().await.unwrap();
-        assert_eq!(expect, got);
-
-        // update store
-        expect
-            .update(StatusRow::new(store_name, 1, ServiceStatus::Unknown))
-            .unwrap();
-        store_reporter
-            .set_service_status(store::SERVICE_NAME, ServingStatus::Unknown)
-            .await;
-        let got = stream.next().await.unwrap();
-        assert_eq!(expect, got);
-
-        gateway_task.abort();
-        p2p_task.abort();
-        store_task.abort();
     }
 }
