+++ iroh-rpc-client/src/gateway.rs
@@ -1,25 +1,53 @@
-#[cfg(feature = "grpc")]
-use crate::status::{self, StatusRow};
+use super::status::StatusRow;
+use super::ServiceStatus;
 use anyhow::Result;
-#[cfg(feature = "grpc")]
 use futures::Stream;
-#[cfg(feature = "grpc")]
-use iroh_rpc_types::gateway::gateway_client::GatewayClient as GrpcGatewayClient;
-use iroh_rpc_types::{
-    gateway::{Gateway, GatewayClientAddr, GatewayClientBackend},
-    Addr,
-};
-#[cfg(feature = "grpc")]
-use tonic::transport::Endpoint;
-#[cfg(feature = "grpc")]
-use tonic_health::proto::health_client::HealthClient;
+use iroh_rpc_types::gateway::*;
+use std::fmt;
 
-impl_client!(Gateway);
+pub(crate) const NAME: &str = "gateway";
+
+#[derive(Clone)]
+pub struct GatewayClient {
+    client: quic_rpc::RpcClient<GatewayService, crate::ChannelTypes>,
+}
+
+impl fmt::Debug for GatewayClient {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("GatewayClient2")
+            .field("client", &self.client)
+            .finish()
+    }
+}
 
 impl GatewayClient {
+    pub async fn new(addr: GatewayAddr) -> anyhow::Result<Self> {
+        let client = crate::open_client::<GatewayService>(addr).await?;
+        Ok(Self { client })
+    }
+
     #[tracing::instrument(skip(self))]
     pub async fn version(&self) -> Result<String> {
-        let res = self.backend.version(()).await?;
+        let res = self.client.rpc(VersionRequest).await?;
         Ok(res.version)
+    }
+
+    #[tracing::instrument(skip(self))]
+    pub async fn check(&self) -> StatusRow {
+        let status: ServiceStatus = self
+            .version()
+            .await
+            .map(|_| ServiceStatus::Serving)
+            .unwrap_or_else(|_e| ServiceStatus::Unknown);
+        StatusRow {
+            name: "gateway",
+            number: 1,
+            status,
+        }
+    }
+
+    #[tracing::instrument(skip(self))]
+    pub async fn watch(&self) -> impl Stream<Item = StatusRow> {
+        futures::stream::pending()
     }
 }
