--- iroh-rpc-types/proto/gateway.proto
@@ -1,13 +1,0 @@
-syntax = "proto3";
-
-package gateway;
-
-import "google/protobuf/empty.proto";
-
-service Gateway {
-  rpc Version(google.protobuf.Empty) returns (VersionResponse) {}
-}
-
-message VersionResponse {
-  string version = 1;
-}
