--- iroh-rpc-types/proto/test.proto
@@ -1,5 +1,0 @@
-syntax = "proto3";
-
-package test;
-
-service Test {}
