+++ iroh-rpc-types/src/addr.rs
@@ -1,101 +1,96 @@
+use anyhow::{anyhow, bail};
+use quic_rpc::{transport::mem, Service};
+use serde_with::{DeserializeFromStr, SerializeDisplay};
 use std::{
     fmt::{Debug, Display},
     net::SocketAddr,
     str::FromStr,
 };
 
-use anyhow::{anyhow, bail};
-use serde_with::{DeserializeFromStr, SerializeDisplay};
-use tokio::sync::mpsc::{Receiver, Sender};
-
-#[derive(SerializeDisplay, DeserializeFromStr, Clone)]
-pub enum Addr<T = ()> {
-    #[cfg(feature = "grpc")]
-    GrpcHttp2(SocketAddr),
-    #[cfg(feature = "grpc")]
-    GrpcHttp2Lookup(String),
-    #[cfg(feature = "mem")]
-    Mem(T),
+/// An address. This can be either a memory address, already containing the channel, or a network
+/// address which will have to be opened.
+#[derive(SerializeDisplay, DeserializeFromStr)]
+pub enum Addr<S: Service> {
+    Irpc(SocketAddr),
+    IrpcLookup(String),
+    Mem(
+        mem::ServerChannel<S::Req, S::Res>,
+        mem::ClientChannel<S::Res, S::Req>,
+    ),
 }
 
-impl<T> PartialEq for Addr<T> {
+impl<S: Service> PartialEq for Addr<S> {
     fn eq(&self, other: &Self) -> bool {
         match (self, other) {
-            #[cfg(feature = "grpc")]
-            (Self::GrpcHttp2(addr1), Self::GrpcHttp2(addr2)) => addr1.eq(addr2),
-            #[cfg(feature = "grpc")]
-            (Self::GrpcHttp2Lookup(addr1), Self::GrpcHttp2Lookup(addr2)) => addr1.eq(addr2),
+            (Self::Irpc(addr1), Self::Irpc(addr2)) => addr1.eq(addr2),
+            (Self::IrpcLookup(addr1), Self::IrpcLookup(addr2)) => addr1.eq(addr2),
             _ => false,
         }
     }
 }
 
-impl<T> Addr<T> {
-    pub fn new_mem() -> (Addr<Receiver<T>>, Addr<Sender<T>>) {
-        let (s, r) = tokio::sync::mpsc::channel(256);
+impl<S: Service> Addr<S> {
+    pub fn new_mem() -> Self {
+        let (server, client) = mem::connection(1);
 
-        (Addr::Mem(r), Addr::Mem(s))
+        Self::Mem(server, client)
     }
 }
 
-impl<T> Addr<T> {
+impl<S: Service> Addr<S> {
     pub fn try_as_socket_addr(&self) -> Option<SocketAddr> {
-        #[cfg(feature = "grpc")]
-        if let Addr::GrpcHttp2(addr) = self {
+        if let Addr::Irpc(addr) = self {
             return Some(*addr);
         }
         None
     }
 }
 
-impl<T> Display for Addr<T> {
+impl<S: Service> Display for Addr<S> {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
-            #[cfg(feature = "grpc")]
-            Addr::GrpcHttp2(addr) => write!(f, "grpc://{}", addr),
-            #[cfg(feature = "grpc")]
-            Addr::GrpcHttp2Lookup(addr) => write!(f, "grpc://{}", addr),
-            #[cfg(feature = "mem")]
-            Addr::Mem(_) => write!(f, "mem"),
+            Addr::Irpc(addr) => write!(f, "irpc://{}", addr),
+            Addr::IrpcLookup(addr) => write!(f, "irpc://{}", addr),
+            Addr::Mem(_, _) => write!(f, "mem"),
             #[allow(unreachable_patterns)]
             _ => unreachable!(),
         }
     }
 }
 
-impl<T> Debug for Addr<T> {
+impl<S: Service> Debug for Addr<S> {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         Display::fmt(self, f)
     }
 }
 
-impl<T> FromStr for Addr<T> {
+impl<S: Service> Clone for Addr<S> {
+    fn clone(&self) -> Self {
+        match self {
+            Addr::Irpc(addr) => Addr::Irpc(*addr),
+            Addr::IrpcLookup(addr) => Addr::IrpcLookup(addr.clone()),
+            Addr::Mem(server, client) => Addr::Mem(server.clone(), client.clone()),
+        }
+    }
+}
+
+impl<S: Service> FromStr for Addr<S> {
     type Err = anyhow::Error;
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
-        #[cfg(feature = "mem")]
         if s == "mem" {
             bail!("memory addresses can not be serialized or deserialized");
         }
 
-        #[cfg(feature = "grpc")]
-        {
-            use std::net::ToSocketAddrs;
-            let mut parts = s.split("://");
-            if let Some(prefix) = parts.next() {
-                if prefix == "grpc" {
-                    if let Some(part) = parts.next() {
-                        if let Ok(addr) = part.parse::<SocketAddr>() {
-                            return Ok(Addr::GrpcHttp2(addr));
-                        }
-                        // attempt to resolve the address, if it can be resolved,
-                        // it's considered a lookup address
-                        if let Ok(mut addr_iter) = part.to_socket_addrs() {
-                            if addr_iter.next().is_some() {
-                                return Ok(Addr::GrpcHttp2Lookup(String::from(part)));
-                            }
-                        }
-                    }
+        let mut parts = s.split("://");
+        if let Some(prefix) = parts.next() {
+            if prefix == "irpc" {
+                if let Some(part) = parts.next() {
+                    return Ok(if let Ok(addr) = part.parse() {
+                        Addr::Irpc(addr)
+                    } else {
+                        Addr::IrpcLookup(part.to_string())
+                    });
                 }
             }
         }
@@ -106,27 +101,17 @@ impl<T> FromStr for Addr<T> {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
 
-    #[cfg(feature = "grpc")]
     #[test]
-    fn test_addr_roundtrip_grpc_http2() {
+    fn test_addr_roundtrip_irpc_http2() {
+        use crate::gateway::GatewayAddr;
+        use crate::Addr;
+        use std::net::SocketAddr;
+
         let socket: SocketAddr = "198.168.2.1:1234".parse().unwrap();
-        let addr = Addr::GrpcHttp2(socket);
+        let addr = Addr::Irpc(socket);
 
-        assert_eq!(addr.to_string().parse::<Addr>().unwrap(), addr);
-        assert_eq!(addr.to_string(), "grpc://198.168.2.1:1234");
-    }
-
-    // TODO(b5): only running this on unix b/c windows doesn't have localhost
-    // enabled by default
-    #[cfg(all(feature = "grpc", unix))]
-    #[test]
-    fn test_addr_roundtrip_http2_lookup() {
-        let name = "localhost:1234".to_string();
-        let addr = Addr::GrpcHttp2Lookup(name);
-
-        assert_eq!(addr.to_string().parse::<Addr>().unwrap(), addr);
-        assert_eq!(addr.to_string(), "grpc://localhost:1234");
+        assert_eq!(addr.to_string().parse::<GatewayAddr>().unwrap(), addr);
+        assert_eq!(addr.to_string(), "irpc://198.168.2.1:1234");
     }
 }
