--- iroh-rpc-types/src/test.rs
@@ -1,1 +1,0 @@
-include_proto!("test");
