+++ iroh-store/Cargo.toml
@@ -21,10 +21,8 @@ ctrlc = "3.2.2"
 futures = "0.3.21"
 git-version = "0.3.5"
 iroh-metrics = { version = "0.1.3", path = "../iroh-metrics", default-features = false, features=["store"] }
-# iroh-rpc-client = { version = "0.1.3", path = "../iroh-rpc-client", default-features = false }
-# iroh-rpc-types = { version = "0.1.3", path = "../iroh-rpc-types", default-features = false }
-iroh-rpc-client = { package = "iroh-qrpc-client", version = "0.2.0", path = "../iroh-qrpc-client", default-features = false }
-iroh-rpc-types = { package = "iroh-qrpc-types", version = "0.2.0", path = "../iroh-qrpc-types", default-features = false }
+iroh-rpc-client = { version = "0.1.3", path = "../iroh-rpc-client", default-features = false }
+iroh-rpc-types = { version = "0.1.3", path = "../iroh-rpc-types", default-features = false }
 iroh-util = { version = "0.1.3", path = "../iroh-util" }
 multihash = "0.17"
 names = { version = "0.14.0", default-features = false }
